Init snapshot tess
