+++ src/index.js
@@ -1,59 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
-
-const accommodations = [
-  {
-    id: 1,
-    description: `Beautiful & luxurious apartment at great location`,
-    type: `Apartment`,
-    price: 120,
-    image: `img/apartment-01.jpg`,
-    rating: 100,
-    isPremium: true,
-    isBookmark: false
-  },
-  {
-    id: 2,
-    description: `Wood and stone place`,
-    type: `Private room`,
-    price: 80,
-    image: `img/room.jpg`,
-    rating: 50,
-    isPremium: false,
-    isBookmark: true,
-  },
-  {
-    id: 3,
-    description: `Canal View Prinsengracht`,
-    type: `Apartment`,
-    price: 132,
-    image: `img/apartment-02.jpg`,
-    rating: 80,
-    isPremium: false,
-    isBookmark: false
-  },
-  {
-    id: 4,
-    description: `Nice, cozy, warm big bed apartment`,
-    type: `Apartment`,
-    price: 180,
-    image: `img/apartment-03.jpg`,
-    rating: 100,
-    isPremium: true,
-    isBookmark: false
-  },
-  {
-    id: 5,
-    description: `Wood and stone place`,
-    type: `Private room`,
-    price: 80,
-    image: `img/room.jpg`,
-    rating: 80,
-    isPremium: false,
-    isBookmark: false
-  },
-];
+import {accommodations} from "./mocks.js";
 
 const init = () => {
   ReactDOM.render(
